+++ src/core/context/context-api.ts
@@ -20,7 +20,12 @@ import { base58 } from '../../util/encoding'
 import { makeAccount } from '../account/account-init'
 import { createLogin, usernameAvailable } from '../login/create'
 import { requestEdgeLogin } from '../login/edge'
-import { makeAuthJson, searchTree, syncLogin } from '../login/login'
+import {
+  decryptChildKey,
+  makeAuthJson,
+  searchTree,
+  syncLogin
+} from '../login/login'
 import { loginFetch } from '../login/login-fetch'
 import { fetchLoginMessages } from '../login/login-messages'
 import {
@@ -48,25 +53,22 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
     // Persist disabled duress mode
     await clientFile.save(ai.props.io.disklet, CLIENT_FILE_NAME, {
       ...ai.props.state.clientInfo,
-      duressLoginId: undefined
+      duressEnabled: false
     })
     // Disable duress mode
     ai.props.dispatch({
       type: 'LOGIN_DURESS_MODE_DISABLED'
     })
   }
-  async function enableDuressMode(duressLoginId: Uint8Array): Promise<void> {
+  async function enableDuressMode(): Promise<void> {
     // Persist enabled duress mode
     await clientFile.save(ai.props.io.disklet, CLIENT_FILE_NAME, {
       ...ai.props.state.clientInfo,
-      duressLoginId
+      duressEnabled: true
     })
     // Enable duress mode
     ai.props.dispatch({
-      type: 'LOGIN_DURESS_MODE_ENABLED',
-      payload: {
-        duressLoginId
-      }
+      type: 'LOGIN_DURESS_MODE_ENABLED'
     })
   }
 
@@ -173,17 +175,35 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
       ai.props.log.breadcrumb('EdgeContext.loginWithPassword', {})
 
       username = fixUsername(username)
-      const stashTree = getStashByUsername(ai, username)
-      const sessionKey = await loginPassword(
-        ai,
-        stashTree ?? getEmptyStash(username),
-        password,
-        opts
-      )
-      // Disable duress mode if it is setup and active
-      if (ai.props.state.clientInfo.duressLoginId != null) {
-        await disableDuressMode()
+      // If we don't have a stash for this username,
+      // then this must be a first-time login on this device:
+      const stash = getStashByUsername(ai, username) ?? getEmptyStash(username)
+      const sessionKey = await loginPassword(ai, stash, password, opts)
+
+      // Attempt to log into duress account if duress mode is enabled:
+      if (ai.props.state.clientInfo.duressEnabled) {
+        const duressAppId = appId + '.duress'
+        const duressStash = searchTree(
+          stash,
+          stash => stash.appId === duressAppId
+        )
+        // We may still be in duress mode but not log in into a duress account
+        // if it does not exist. It's important that we do not disable duress
+        // mode from this routine to make sure other accounts with duress mode
+        // still are protected.
+        if (duressStash != null) {
+          const duressSessionKey = decryptChildKey(
+            stash,
+            sessionKey,
+            duressStash.loginId
+          )
+          return await makeAccount(ai, duressSessionKey, 'passwordLogin', {
+            ...opts,
+            duressMode: true
+          })
+        }
       }
+
       return await makeAccount(ai, sessionKey, 'passwordLogin', opts)
     },
 
@@ -224,8 +244,6 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
         mainStash: LoginStash
       ): Promise<EdgeAccount> {
         const sessionKey = await loginPin2(ai, stashTree, mainStash, pin, opts)
-        // Disable duress mode if it is setup and active
-        await disableDuressMode()
         // Make the account for the main account
         return await makeAccount(ai, sessionKey, 'pinLogin', opts)
       }
@@ -242,8 +260,6 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
           pin,
           opts
         )
-        // The original account will be used for display
-        await enableDuressMode(stashTree.loginId)
         // Make the account with duress mode enabled
         return await makeAccount(ai, sessionKey, 'pinLogin', {
           ...opts,
@@ -253,19 +269,18 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
 
       // No duress account configured, so just login to the main account:
       if (duressStash == null) {
+        // It's important that we don't disable duress mode here because
+        // we want to protect account that have duress mode enabled and only
+        // allow those accounts to suspend duress mode.
         return await loginMainAccount(stashTree, mainStash)
       }
 
-      // Check if we are in duress mode for this account:
-      const inDuressModeForAccount =
-        verifyData(
-          ai.props.state.clientInfo.duressLoginId ?? new Uint8Array(),
-          stashTree.loginId
-        ) ?? false
+      // Check if we are in duress mode:
+      const inDuressMode = ai.props.state.clientInfo.duressEnabled
 
       // Try pin-login on either the duress or main accounts, smartly:
       try {
-        return inDuressModeForAccount
+        return inDuressMode
           ? await loginDuressAccount(stashTree, duressStash)
           : await loginMainAccount(stashTree, mainStash)
       } catch (error) {
@@ -276,9 +291,16 @@ export function makeContextApi(ai: ApiInput): EdgeContext {
         ) {
           throw error
         }
-        return inDuressModeForAccount
+        const account = inDuressMode
           ? await loginMainAccount(stashTree, mainStash)
           : await loginDuressAccount(stashTree, duressStash)
+        // Only Enable/Disable duress mode if account creation was success.
+        if (inDuressMode) {
+          await disableDuressMode()
+        } else {
+          await enableDuressMode()
+        }
+        return account
       }
     },